requirements.txt
pyperclip
paho-mqtt
Pillow
xkcdpass
cryptography